requirements.txt
backcall==0.1.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
ipdb==0.12.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.1
parso==0.5.2
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.2
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
six==1.14.0
traitlets==4.3.3
wcwidth==0.1.8